requirements.txt
moviepy>=1.0.3,<2.0
pyannote.audio>=3.1.1
python-dotenv>=1.0.0
pytest>=7.0.0
rembg>=2.0.57
Pillow>=10.0.0
imageio>=2.31.0
numpy>=1.24.0,<2.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# IMPORTANT: Install PyTorch appropriate for your platform separately if needed.
# On Apple Silicon (MPS), visit https://pytorch.org/get-started/locally/
# Example (as of 2025, may change):
# pip install torch torchvision torchaudio


